+++ src/util/path.ts
@@ -2,11 +2,11 @@ const proxy = null;
 const isDev = process.env.NODE_ENV === 'development';
 
 const prod = {
-  upload: '../../../SysManage/AjaxHandler/UploadHandler.ashx', // 主程序的上传接口
+  upload: '../../../SysManage/AjaxHandler/UploadHandler.ashx',
 };
 
 const dev = {
-  upload: '../../../SysManage/AjaxHandler/UploadHandler.ashx', // 主程序的上传接口
+  upload: '../../../SysManage/AjaxHandler/UploadHandler.ashx',
 };
 
 const path = isDev ? dev : prod;
